dff_Artificial_Inteligence_skill/requirements.txt
df_engine==0.8.1
nltk==3.5
pyyaml==6.0
ruamel.yaml==0.17.20